Update Docker build process and introduce multi-stage build
This commit modifies the Docker run command in the start-up script to use a newer version of the Gradle image along with JDK 11. It also introduces a new Dockerfile for a multi-stage build process, where the application is initially built using a Gradle image, and the resulting jar file is then copied over to a new OpenJDK image for final run, improving overall efficiency and reducing final image size.